Add some more docs
