+++ src/action/common/configure_nix.rs
@@ -1,7 +1,8 @@
 use crate::{
     action::{
-        base::{ConfigureNixDaemonService, SetupDefaultProfile},
+        base::SetupDefaultProfile,
         common::{ConfigureShellProfile, PlaceChannelConfiguration, PlaceNixConfiguration},
+        linux::ConfigureNixDaemonService,
         Action, ActionDescription, StatefulAction,
     },
     channel_value::ChannelValue,
@@ -11,6 +12,9 @@ use crate::{
 
 use reqwest::Url;
 
+/**
+Configure Nix and start it
+ */
 #[derive(Debug, serde::Deserialize, serde::Serialize, Clone)]
 pub struct ConfigureNix {
     setup_default_profile: StatefulAction<SetupDefaultProfile>,
