+++ src/action/common/mod.rs
@@ -1,9 +1,9 @@
-/*! Actions which only call other base plugins. */
+//! [`Action`](crate::action::Action)s which only call other base plugins
 
 mod configure_nix;
 mod configure_shell_profile;
 mod create_nix_tree;
-mod create_users_and_group;
+mod create_users_and_groups;
 mod place_channel_configuration;
 mod place_nix_configuration;
 mod provision_nix;
@@ -11,7 +11,7 @@ mod provision_nix;
 pub use configure_nix::ConfigureNix;
 pub use configure_shell_profile::ConfigureShellProfile;
 pub use create_nix_tree::{CreateNixTree, CreateNixTreeError};
-pub use create_users_and_group::{CreateUsersAndGroup, CreateUsersAndGroupError};
+pub use create_users_and_groups::{CreateUsersAndGroups, CreateUsersAndGroupsError};
 pub use place_channel_configuration::{PlaceChannelConfiguration, PlaceChannelConfigurationError};
 pub use place_nix_configuration::{PlaceNixConfiguration, PlaceNixConfigurationError};
 pub use provision_nix::{ProvisionNix, ProvisionNixError};
