+++ src/action/darwin/mod.rs
@@ -1,17 +1,20 @@
-mod bootstrap_volume;
-mod create_apfs_volume;
-mod create_synthetic_objects;
-mod create_volume;
-mod enable_ownership;
-mod encrypt_volume;
-mod kickstart_launchctl_service;
-mod unmount_volume;
+/*!  [`Action`](crate::action::Action)s for Darwin based systems
+*/
 
-pub use bootstrap_volume::{BootstrapVolume, BootstrapVolumeError};
-pub use create_apfs_volume::{CreateApfsVolume, CreateApfsVolumeError, NIX_VOLUME_MOUNTD_DEST};
+mod bootstrap_apfs_volume;
+mod create_apfs_volume;
+mod create_nix_volume;
+mod create_synthetic_objects;
+mod enable_ownership;
+mod encrypt_apfs_volume;
+mod kickstart_launchctl_service;
+mod unmount_apfs_volume;
+
+pub use bootstrap_apfs_volume::{BootstrapApfsVolume, BootstrapVolumeError};
+pub use create_apfs_volume::{CreateApfsVolume, CreateVolumeError};
+pub use create_nix_volume::{CreateApfsVolumeError, CreateNixVolume, NIX_VOLUME_MOUNTD_DEST};
 pub use create_synthetic_objects::{CreateSyntheticObjects, CreateSyntheticObjectsError};
-pub use create_volume::{CreateVolume, CreateVolumeError};
 pub use enable_ownership::{EnableOwnership, EnableOwnershipError};
-pub use encrypt_volume::{EncryptVolume, EncryptVolumeError};
+pub use encrypt_apfs_volume::{EncryptApfsVolume, EncryptVolumeError};
 pub use kickstart_launchctl_service::{KickstartLaunchctlService, KickstartLaunchctlServiceError};
-pub use unmount_volume::{UnmountVolume, UnmountVolumeError};
+pub use unmount_apfs_volume::{UnmountApfsVolume, UnmountVolumeError};
