+++ src/action/linux/mod.rs
@@ -1,5 +1,9 @@
+//!  [`Action`](crate::action::Action)s for Linux based systems
+
+mod configure_nix_daemon_service;
 mod create_systemd_sysext;
 mod start_systemd_unit;
 
+pub use configure_nix_daemon_service::{ConfigureNixDaemonService, ConfigureNixDaemonServiceError};
 pub use create_systemd_sysext::{CreateSystemdSysext, CreateSystemdSysextError};
 pub use start_systemd_unit::{StartSystemdUnit, StartSystemdUnitError};
