+++ src/planner/linux/mod.rs
@@ -1,3 +1,5 @@
+//! Planners for Linux based systems
+
 mod multi;
 
 pub use multi::LinuxMulti;
